Fix a bug ...

+++ DirectoryTraversal.h
@@ -65,7 +65,7 @@ public:
 					continue;
 
 				if (dirEntry.is_regular_file()) {
-					if (file_size(dirEntry.path()) < m_refConfig.m_minFileSize)
+					if (static_cast<unsigned int>(file_size(dirEntry.path())) < m_refConfig.m_minFileSize)
 						continue;
 					
 						if (std::regex_search(directoryPath, regularPattern)) {
